+++ noodles-bam/examples/bam_filter_data.rs
@@ -24,7 +24,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = args.next().expect("missing src");
 
     let mut reader = bam::reader::Builder::default().build_from_path(src)?;
-    let header = reader.read_header()?.parse()?;
+    let header = reader.read_header()?;
     reader.read_reference_sequences()?;
 
     let stdout = io::stdout().lock();