+++ noodles-bam/examples/bam_filter_read_name.rs
@@ -54,7 +54,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let read_names = read_read_names(read_names_src)?;
 
     let mut reader = bam::reader::Builder::default().build_from_path(src)?;
-    let header = reader.read_header()?.parse()?;
+    let header = reader.read_header()?;
     reader.read_reference_sequences()?;
 
     let stdout = io::stdout().lock();