+++ noodles-bam/examples/bam_reset.rs
@@ -20,7 +20,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = bam::reader::Builder::default().build_from_path(src)?;
-    let header: sam::Header = reader.read_header()?.parse()?;
+    let header = reader.read_header()?;
     reader.read_reference_sequences()?;
 
     let stdout = BufWriter::new(io::stdout().lock());